Move context into own folder
