+++ src/state/Calculator/Context.tsx
@@ -1,23 +1,15 @@
 import experienceTable from "data/experienceTable";
 import { createContext, Dispatch, ReactNode, useReducer } from "react";
+import { CalculatorAction, CalculatorActionTypes } from "./actions";
 
-enum ActionTypes {
-  SET_XP = "SET_XP",
-  SET_PLAYER_MASTERY = "SET_PLAYER_MASTERY",
-}
 type State = {
   currentExp: string;
   playerMastery: string;
 };
-type SetXPAction = { type: ActionTypes.SET_XP; payload: string };
-type SetPlayerMasteryAction = {
-  type: ActionTypes.SET_PLAYER_MASTERY;
-  payload: string;
-};
-type Action = SetXPAction | SetPlayerMasteryAction;
+
 type CalculatorContextProps = {
   calculatorState: State;
-  calculatorDispatch: Dispatch<Action>;
+  calculatorDispatch: Dispatch<CalculatorAction>;
 };
 type CalculatorProviderProps = {
   children: ReactNode;
@@ -28,16 +20,16 @@ const initialState: State = {
   playerMastery: "1",
 };
 
-const CalculatorReducer = (state: State, action: Action): State => {
+const CalculatorReducer = (state: State, action: CalculatorAction): State => {
   switch (action.type) {
-    case ActionTypes.SET_XP: {
+    case CalculatorActionTypes.SET_XP: {
       const { payload } = action;
       if (Number(payload) < 0) return { ...state, currentExp: "0" };
       if (Number(payload) > experienceTable[98])
         return { ...state, currentExp: experienceTable[98].toString() };
       return { ...state, currentExp: payload };
     }
-    case ActionTypes.SET_PLAYER_MASTERY:
+    case CalculatorActionTypes.SET_PLAYER_MASTERY:
       return { ...state, playerMastery: action.payload };
     default:
       return { ...state };
@@ -67,12 +59,3 @@ export const CalculatorProvider = (
     </CalculatorContext.Provider>
   );
 };
-
-export const setXp = (xp: string): SetXPAction => ({
-  type: ActionTypes.SET_XP,
-  payload: xp,
-});
-export const setPlayerMastery = (mastery: string): SetPlayerMasteryAction => ({
-  type: ActionTypes.SET_PLAYER_MASTERY,
-  payload: mastery,
-});
